+++ Src/DependencyCollector/Shared.Tests/HeaderCollectionManupilationTests.cs
@@ -1,9 +1,8 @@
 ﻿namespace Microsoft.ApplicationInsights.DependencyCollector
 {
-    using System;
     using System.Collections.Generic;
+    using System.Linq;
     using System.Net;
-    using System.Text;
     using Common;
     using VisualStudio.TestTools.UnitTesting;
 
@@ -59,13 +58,108 @@
             // Non empty collection - adding new key
             headers.SetNameValueHeaderValue("Request-Context", "roleName", "workerRole");
             Assert.AreEqual(1, headers.Keys.Count);
-            Assert.AreEqual("appId=appIdValue, roleName=workerRole", headers["Request-Context"]);
+            Assert.AreEqual("appId=appIdValue,roleName=workerRole", headers["Request-Context"]);
 
             // overwritting existing key
             headers.SetNameValueHeaderValue("Request-Context", "roleName", "webRole");
             headers.SetNameValueHeaderValue("Request-Context", "appId", "udpatedAppId");
             Assert.AreEqual(1, headers.Keys.Count);
-            Assert.AreEqual("appId=udpatedAppId, roleName=webRole", headers["Request-Context"]);
+            Assert.AreEqual("appId=udpatedAppId,roleName=webRole", headers["Request-Context"]);
+        }
+
+        /// <summary>
+        /// Tests that GetCollectionFromHeaderNoDuplicates gets collection of key-value pairs from existing, non-empty header.
+        /// </summary>
+        [TestMethod]
+        public void GetCollectionFromHeaderNoDuplicates()
+        {
+            WebHeaderCollection headers = new WebHeaderCollection();
+            headers["Correlation-Context"] = "k1=v1, k2=v2, k3 = v3 ";
+
+            var correlationContext = headers.GetNameValueCollectionFromHeader("Correlation-Context");
+            Assert.IsNotNull(correlationContext);
+            var corrContextDict = correlationContext.ToDictionary(kvp => kvp.Key, kvp => kvp.Value);
+            Assert.AreEqual(3, corrContextDict.Count);
+            Assert.AreEqual("v1", corrContextDict["k1"]);
+            Assert.AreEqual("v2", corrContextDict["k2"]);
+            Assert.AreEqual("v3", corrContextDict["k3"]);
+        }
+
+        /// <summary>
+        /// Tests that GetCollectionFromHeaderNoDuplicates returns null if header does not exists or is empty.
+        /// </summary>
+        [TestMethod]
+        public void GetCollectionFromEmptyHeader()
+        {
+            WebHeaderCollection headers = new WebHeaderCollection();
+            Assert.IsNull(headers.GetNameValueCollectionFromHeader("Correlation-Context"));
+
+            headers["Correlation-Context"] = "   ";
+            var correlationContext = headers.GetNameValueCollectionFromHeader("Correlation-Context");
+            Assert.IsTrue(correlationContext == null || !correlationContext.Any());
+        }
+
+        /// <summary>
+        /// Tests that GetCollectionFromHeaderNoDuplicates gets collection of key-value pairs from existing, non-empty header with duplicates.
+        /// </summary>
+        [TestMethod]
+        public void GetCollectionFromHeaderWithDuplicates()
+        {
+            WebHeaderCollection headers = new WebHeaderCollection();
+            headers["Correlation-Context"] = "k1=v1, k2=v2, k1 = v3";
+
+            var correlationContext = headers.GetNameValueCollectionFromHeader("Correlation-Context").ToArray();
+            Assert.AreEqual(2, correlationContext.Length);
+            Assert.IsTrue(correlationContext.Contains(new KeyValuePair<string, string>("k1", "v1")));
+            Assert.IsTrue(correlationContext.Contains(new KeyValuePair<string, string>("k2", "v2")));
+        }
+
+        /// <summary>
+        /// Tests that GetCollectionFromHeaderNoDuplicates gets collection of key-value pairs from existing
+        /// non-empty invalid header.
+        /// </summary>
+        [TestMethod]
+        public void GetCollectionFromInvalidHeader()
+        {
+            WebHeaderCollection headers = new WebHeaderCollection();
+
+            // no valid items
+            headers["Correlation-Context"] = "k1, some string,k2=v2=v3";
+
+            var correlationContext = headers.GetNameValueCollectionFromHeader("Correlation-Context");
+            Assert.IsTrue(correlationContext == null || !correlationContext.Any());
+
+            // some valid items
+            headers["Correlation-Context"] = "k1=v1, some string";
+
+            correlationContext = headers.GetNameValueCollectionFromHeader("Correlation-Context");
+            Assert.IsNotNull(correlationContext);
+            Assert.AreEqual(1, correlationContext.Count());
+            Assert.IsTrue(correlationContext.Contains(new KeyValuePair<string, string>("k1", "v1")));
+        }
+
+        [TestMethod]
+        public void SetNameValueHeaderWithEmptyCollectionSetsNothing()
+        {
+            WebHeaderCollection headers = new WebHeaderCollection();
+            headers.SetHeaderFromNameValueCollection("Correlation-Context", new List<KeyValuePair<string, string>>());
+            Assert.IsNull(headers["Correlation-Context"]);
+        }
+
+        [TestMethod]
+        public void SetNameValueHeaderWithNonEmptyCollectionSetsHeader()
+        {
+            WebHeaderCollection headers = new WebHeaderCollection();
+            headers.SetHeaderFromNameValueCollection(
+                "Correlation-Context",
+                new List<KeyValuePair<string, string>>
+                {
+                    new KeyValuePair<string, string>("k1 ", "v1"),
+                    new KeyValuePair<string, string>("k2", " v2"),
+                    new KeyValuePair<string, string>("k1", "v3")
+                });
+            Assert.IsNotNull(headers["Correlation-Context"]);
+            Assert.AreEqual("k1=v1,k2=v2,k1=v3", headers["Correlation-Context"]);
         }
     }
 }