+++ Src/Web/Web.Shared.Net.Tests/ExceptionTrackingTelemetryModuleTests.cs
@@ -31,6 +31,12 @@
             };
         }
 
+        [TestCleanup]
+        public void Cleanup()
+        {
+            Common.ActivityHelpers.StopRequestActivity();
+        }
+
         [TestMethod]
         public void OnErrorTracksExceptionsThatArePresentInHttpContext()
         {