+++ Src/Web/Web.Shared.Net.Tests/HttpContextExtensionTest.cs
@@ -10,6 +10,12 @@
     [TestClass]
     public class HttpContextExtensionTest
     {
+        [TestCleanup]
+        public void Cleanup()
+        {
+            Common.ActivityHelpers.StopRequestActivity();
+        }
+
         [TestMethod]
         public void GetRequestTelemetryReturnsNullForNullContext()
         {