+++ Src/Web/Web.Shared.Net.Tests/OperationCorrelationTelemetryInitializerTests.cs
@@ -1,17 +1,22 @@
 ﻿namespace Microsoft.ApplicationInsights.Web
 {
-    using System;
     using System.Collections.Generic;
-    using System.Web;    
-    using Common;
+    using System.Web;
+    using Microsoft.ApplicationInsights.Common;
     using Microsoft.ApplicationInsights.DataContracts;
     using Microsoft.ApplicationInsights.Web.Helpers;
     using Microsoft.ApplicationInsights.Web.Implementation;
-    using Microsoft.VisualStudio.TestTools.UnitTesting;    
-    
+    using Microsoft.VisualStudio.TestTools.UnitTesting;
+
     [TestClass]
     public class OperationCorrelationTelemetryInitializerTests
     {
+        [TestCleanup]
+        public void Cleanup()
+        {
+            ActivityHelpers.StopRequestActivity();
+        }
+
         [TestMethod]
         public void InitializeDoesNotThrowWhenHttpContextIsNull()
         {
@@ -20,118 +25,83 @@
         }
 
         [TestMethod]
-        public void InitializeSetsParentIdForTelemetryUsingIdFromRequestTelemetry()
+        public void DefaultHeadersOperationCorrelationTelemetryInitializerAreSet()
         {
-            var exceptionTelemetry = new ExceptionTelemetry();
-            var source = new TestableOperationCorrelationTelemetryInitializer(null);
-            var requestTelemetry = source.FakeContext.CreateRequestTelemetryPrivate();
-
-            source.Initialize(exceptionTelemetry);
-
-            Assert.AreEqual(requestTelemetry.Id, exceptionTelemetry.Context.Operation.ParentId);
+            var initializer = new OperationCorrelationTelemetryInitializer();
+            Assert.AreEqual(RequestResponseHeaders.StandardParentIdHeader, initializer.ParentOperationIdHeaderName);
+            Assert.AreEqual(RequestResponseHeaders.StandardRootIdHeader, initializer.RootOperationIdHeaderName);
         }
 
         [TestMethod]
-        public void InitializeDoesNotOverrideCustomerParentOperationId()
+        public void CustomHeadersOperationCorrelationTelemetryInitializerAreSetProperly()
         {
-            var source = new TestableOperationCorrelationTelemetryInitializer(null);
+            var initializer = new OperationCorrelationTelemetryInitializer();
+            initializer.ParentOperationIdHeaderName = "myParentHeader";
+            initializer.RootOperationIdHeaderName = "myRootHeader";
 
-            var customerTelemetry = new TraceTelemetry("Text");
-            customerTelemetry.Context.Operation.ParentId = "CustomId";
+            Assert.AreEqual("myParentHeader", ActivityHelpers.ParentOperationIdHeaderName);
+            Assert.AreEqual("myRootHeader", ActivityHelpers.RootOperationIdHeaderName);
 
-            source.Initialize(customerTelemetry);
-
-            Assert.AreEqual("CustomId", customerTelemetry.Context.Operation.ParentId);
+            Assert.AreEqual("myParentHeader", initializer.ParentOperationIdHeaderName);
+            Assert.AreEqual("myRootHeader", initializer.RootOperationIdHeaderName);
         }
 
         [TestMethod]
-        public void InitializeSetsRootOperationIdForTelemetryUsingIdFromRequestTelemetry()
+        public void OperationContextIsSetForNonRequestTelemetry()
         {
-            var exceptionTelemetry = new ExceptionTelemetry();
-            var source = new TestableOperationCorrelationTelemetryInitializer(null);
-            var requestTelemetry = source.FakeContext.CreateRequestTelemetryPrivate();
-            requestTelemetry.Context.Operation.Id = "RootId";
+            var source = new TestableOperationCorrelationTelemetryInitializer(new Dictionary<string, string>
+            {
+                ["Request-Id"] = "|guid.1",
+                ["Correlation-Context"] = "k1=v1,k2=v2,k1=v3"
+            });
 
+            // simulate OnBegin behavior:
+            // create telemetry and start activity for children
+            var requestTelemetry = source.FakeContext.CreateRequestTelemetryPrivate();
+            
+            // lost Acitivity / call context
+            ActivityHelpers.StopRequestActivity();
+
+            var exceptionTelemetry = new ExceptionTelemetry();
             source.Initialize(exceptionTelemetry);
 
             Assert.AreEqual(requestTelemetry.Context.Operation.Id, exceptionTelemetry.Context.Operation.Id);
+            Assert.AreEqual(requestTelemetry.Id, exceptionTelemetry.Context.Operation.ParentId);
+
+            Assert.AreEqual(2, exceptionTelemetry.Context.Properties.Count);
+            Assert.AreEqual("v1", exceptionTelemetry.Context.Properties["k1"]);
+            Assert.AreEqual("v2", exceptionTelemetry.Context.Properties["k2"]);
         }
 
         [TestMethod]
-        public void InitializeDoesNotOverrideCustomerRootOperationId()
+        public void OperationContextIsNotUpdatedIfOperationIdIsSet()
         {
-            var source = new TestableOperationCorrelationTelemetryInitializer(null);
-            var requestTelemetry = source.FakeContext.CreateRequestTelemetryPrivate();
-            requestTelemetry.Context.Operation.Id = "RootId";
+            var source = new TestableOperationCorrelationTelemetryInitializer(new Dictionary<string, string>
+            {
+                ["Request-Id"] = "|guid.1",
+                ["Correaltion-Context"] = "k1=v1"
+            });
 
-            var customerTelemetry = new TraceTelemetry("Text");
-            customerTelemetry.Context.Operation.Id = "CustomId";
+            // create telemetry and immediately clean call context/activity
+            source.FakeContext.CreateRequestTelemetryPrivate();
+            ActivityHelpers.StopRequestActivity();
 
-            source.Initialize(customerTelemetry);
+            var exceptionTelemetry = new ExceptionTelemetry();
+            exceptionTelemetry.Context.Operation.Id = "guid";
+            source.Initialize(exceptionTelemetry);
 
-            Assert.AreEqual("CustomId", customerTelemetry.Context.Operation.Id);
-        }
+            Assert.IsNull(exceptionTelemetry.Context.Operation.ParentId);
 
-        [TestMethod]
-        public void InitializeSetsRequestTelemetryRootOperaitonIdToOepraitonId()
-        {
-            var source = new TestableOperationCorrelationTelemetryInitializer(null);
-            var requestTelemetry = source.FakeContext.CreateRequestTelemetryPrivate();
-
-            var customerTelemetry = new TraceTelemetry("Text");
-
-            source.Initialize(customerTelemetry);
-
-            Assert.AreEqual(requestTelemetry.Id, requestTelemetry.Context.Operation.Id);
-        }
-
-        [TestMethod]
-        public void InitializeReadsParentIdFromCustomHeader()
-        {
-            var source = new TestableOperationCorrelationTelemetryInitializer(new Dictionary<string, string>() { { "headerName", "ParentId" } });
-            source.ParentOperationIdHeaderName = "headerName";
-
-            var customerTelemetry = new TraceTelemetry("Text");
-
-            source.Initialize(customerTelemetry);
-
-            var requestTelemetry = source.FakeContext.ReadOrCreateRequestTelemetryPrivate();
-            Assert.AreEqual("ParentId", requestTelemetry.Context.Operation.ParentId);
-        }
-
-        [TestMethod]
-        public void InitializeReadsRootIdFromCustomHeader()
-        {
-            var source = new TestableOperationCorrelationTelemetryInitializer(new Dictionary<string, string>() { { "headerName", "RootId" } });
-            source.RootOperationIdHeaderName = "headerName";
-
-            var customerTelemetry = new TraceTelemetry("Text");
-
-            source.Initialize(customerTelemetry);
-            Assert.AreEqual("RootId", customerTelemetry.Context.Operation.Id);
-
-            var requestTelemetry = source.FakeContext.ReadOrCreateRequestTelemetryPrivate();
-            Assert.AreEqual("RootId", requestTelemetry.Context.Operation.Id);
-        }
-
-        [TestMethod]
-        public void InitializeDoNotMakeRequestAParentOfItself()
-        {
-            var source = new TestableOperationCorrelationTelemetryInitializer(null);
-            var requestTelemetry = source.FakeContext.ReadOrCreateRequestTelemetryPrivate();
-
-            source.Initialize(requestTelemetry);
-            Assert.AreEqual(null, requestTelemetry.Context.Operation.ParentId);
-            Assert.AreEqual(requestTelemetry.Id, requestTelemetry.Context.Operation.Id);
+            Assert.AreEqual(0, exceptionTelemetry.Context.Properties.Count);
         }
 
         private class TestableOperationCorrelationTelemetryInitializer : OperationCorrelationTelemetryInitializer
         {
             private readonly HttpContext fakeContext;
 
-            public TestableOperationCorrelationTelemetryInitializer(IDictionary<string, string> headers)
+            public TestableOperationCorrelationTelemetryInitializer(IDictionary<string, string> headers = null)
             {
-                 this.fakeContext = HttpModuleHelper.GetFakeHttpContext(headers);
+                this.fakeContext = HttpModuleHelper.GetFakeHttpContext(headers);
             }
 
             public HttpContext FakeContext