+++ Src/Web/Web.Shared.Net.Tests/OperationNameTelemetryInitializerTests.cs
@@ -10,6 +10,12 @@
     [TestClass]
     public class OperationNameTelemetryInitializerTests
     {
+        [TestCleanup]
+        public void Cleanup()
+        {
+           Common.ActivityHelpers.StopRequestActivity();
+        }
+
         [TestMethod]
         public void InitializeDoesNotThrowWhenHttpContextIsNull()
         {