+++ Src/Web/Web.Shared.Net.Tests/SessionTelemetryInitializerTest.cs
@@ -11,6 +11,12 @@
     [TestClass]
     public class SessionTelemetryInitializerTest
     {
+        [TestCleanup]
+        public void Cleanup()
+        {
+            Common.ActivityHelpers.StopRequestActivity();
+        }
+
         [TestMethod]
         public void InitializeDoesNotThrowWhenHttpContextIsNull()
         {