+++ Src/Web/Web.Shared.Net/ApplicationInsightsHttpModule.cs
@@ -69,6 +69,7 @@
                 {
                     context.BeginRequest += this.OnBeginRequest;
                     context.EndRequest += this.OnEndRequest;
+                    context.PreRequestHandlerExecute += this.OnPreRequestHandlerExecute;
                 }
                 catch (Exception exc)
                 {
@@ -149,6 +150,18 @@
             catch (Exception ex)
             {
                 WebEventSource.Log.HookAddOnSendingHeadersFailedWarning(ex.ToInvariantString());
+            }
+        }
+
+        private void OnPreRequestHandlerExecute(object sender, EventArgs eventArgs)
+        {
+            if (this.isEnabled)
+            {
+                HttpApplication httpApplication = (HttpApplication)sender;
+
+                this.TraceCallback("OnPreRequestHandlerExecute", httpApplication);
+
+                this.requestModule?.OnPreRequestHandlerExecute(httpApplication.Context);
             }
         }